+++ src/ast/attachpoint_parser.cpp
@@ -351,7 +351,8 @@ AttachPointParser::State AttachPointParser::uprobe_parser(bool allow_offset,
   {
     // For PID, the target may be skipped
     parts_.push_back(parts_[1]);
-    parts_[1] = "";
+    auto target = get_pid_exe(bpftrace_.pid());
+    parts_[1] = path_for_pid_mountns(bpftrace_.pid(), target);
   }
   if (parts_.size() != 3)
   {
@@ -380,13 +381,7 @@ AttachPointParser::State AttachPointParser::uprobe_parser(bool allow_offset,
 
   if (ap_->target.empty())
   {
-    if (bpftrace_.pid() > 0)
-    {
-      ap_->target = get_pid_exe(bpftrace_.pid());
-      ap_->target = path_for_pid_mountns(bpftrace_.pid(), ap_->target);
-    }
-    else
-      ap_->target = parts_[1];
+    ap_->target = parts_[1];
   }
 
   // Handle uprobe:/lib/asdf:func+0x100 case